Initial commit of the Stacky stack, a generic stack implementation in C.
This initial implementation is relatively primitive and has several issues with it. These will be addressed in future commits.

Most notable issues are the instances of memory leaks due to how data is popped off of the stack, will likely require some modifications on how data is managed in the stack. The other major issue is the lacking of support for floats and doubles, this will likely be addressed sometime in the future.