Added test cases for the Queue module. Most of these tests are not indicative yet, as the code for the Queue module has not actually been written yet.
Added new test cases to both the Linky and Stacky modules. Modified older Stacky tests to use a more iterative (and cleaner) approach to testing. However, this revealed an issue with how the macro expansion occurs for these structures. It would appear that the macro functions cannot be used without surrounding curly brackets in single line statements (if or while loop without a body) without casuing unexpected behaviour. This is because if the macro is not surrounded by brackets, only the first statement of the macro will execute in the condition. However, the flip side of this is with the surrounding brackets, the functions cannot be used to initialize a variable. More effort will need to be made to find a better solution to this issue.