The big Queue update!
Modified Queue unit tests to be more inline with the standard set out by both the Linky and Stacky modules.

Implemented various functions enabling the Queue data structure to be functional. This includes 'Enqueue', 'Dequeue', 'Front', 'Rear' and 'GetSize' functions.

Modified the macro-functions to be better representative of the various functions layed out before.

One thing to note, both the Stacky and the Queue data structures will inheritly have a memory leak. This will likely be an issue with the Linky data structure unit tests aswell. Currently it was assumed that the user would be responsible for deallocating the assigned memory, but this cannot be a viable assumption for both the Stacky and Queue data structures. This is because the nodes are released from memory once they are popped or dequeued respectively. This will have to be looked into.