Updated to 0.2.1
Updated README file to be more relevant to the current version of the project.

Added documentation for the Stacky, Queue, and Binary Tree modules. Updated the documentation for the Linky module.

Added a VersionLog file that will keep track of changes made in various versions. This should effectively be a more concise version of these commit logs, but should focus on more of the user-end information.