v0.2.4
Modified the Queue and Binary Tree modules's 'getSize' functions to be more resistant to failure.

They have been extended from macro-functions to full on functions which will ensure that the given data structure is indeed valid before attempting to dereference it. This should lead to less accidental segmentation faults when the given data structure is empty.

Made modifications to the version log so that smaller version changes use a smaller header than that of larger version changes. This is much more intuitive and just makes more sense.