v0.2.42
Seperated the compilation of the unit tests from the make 'all' target. The make 'all' target will only compile the mdata .so library from now on, allowing for compilation and use of the library without the Check unit-test framework.

Added a new make target, the make 'mtests' target. This target will compile all unit tests. This target should be run before the make 'test' target or ctest are run