v0.2.5
Modified the project's cmake files to make better use of the functionality provided by cmake. Also made modifications to allow for modular compilation of the library. This means that if a user is only looking to use the Linky module of the library, they can specify that be the only module compiled into the library's final build.

Updated documentation to include more in-depth instructions as to how to make use of the new modularity added in this patch. Also fixed some mild mistakes in the Linky documentation.