v0.2.6.1
Some small modifications were made to the previously implemented valgrind tests. The tests will no longer treat "reachable" memory leaks as an error while conducting the valgrind tests. This is for two reasons, the first being that valgrind does not by default consider them to be an error by default and secondthis creates issues when trying to make use of the "fork" mode while using Check. Removed the setting of the 'CK_FORK' environment variable as it is no longer required and didn't actually do anything useful.

The Linky, Stacky, and Binary Tree modules now all pass their respective valgrind tests. The Queue module will be looked at further in a future update.

Modified the 'VersionLog.md' so that some previous tweaks are represented as such.