v0.2.6.2
Fixed a mistake from the v0.2.6.1 tweak where many tests from the Linky module were commented out. This was done for testing purposes, but should not have stayed in the v0.2.6.1 commit. These tests have been un-commented out and the Linky module still passes all tests

Modified the v0.2.6.1 header in the 'VersionLog.md' file so that it uses the appropriate header size